refactor: modularize runtime errors
